requirement.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
django-celery-beat==2.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
idna==3.6
kombu==5.3.4
pillow==11.0.0
prompt-toolkit==3.0.41
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycryptodome==3.18.0
python-dateutil==2.8.2
requests==2.31.0
setuptools==75.2.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
wheel==0.44.0
uwsgi
django-allauth
django-allauth[socialaccount]
WeasyPrint==64.0 
